distsqlrun: move NoopProcessor to its own file
Release note: None
